bridges/adobe-sign/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pdfminer.six>=20231228
pymupdf>=1.24.0
pytesseract>=0.3.10
Pillow>=10.0.0
